renamed header to navbar
